On se passe de type. Champ directement relié à Regle. CRUD en cours d'écriture sur Champ. TODO : reste du CRUD

+++ src/model/regle.go
@@ -1,7 +1,7 @@
 package model
 
 type Regle struct {
-	Id         int64       `json:"id"`
+	Id         int         `json:"id"`
 	Nom        string      `json:"nom"`
 	Parametres []Parametre `json:"parametres"`
 }
